refactor(realtime): the exchange predicate belongs with the status
The predicate is a pure reading of REALTIME_STATUS, and the turn-taking
policy that consults it is leaving the renderer for @sidecar/voice, which
cannot reach the desktop's shared messages.
