+++ apps/desktop/src/shared/bridge.ts
@@ -39,7 +39,7 @@ import {
 import { type AppGuideSnapshot, isAppGuideSnapshot } from "@sidecar/guide";
 import type { RealtimeConnection } from "@sidecar/hosted";
 import type { SupersetSignInSnapshot } from "@sidecar/providers/superset/sign-in-stage";
-import type { RealtimeDiagnostics } from "@sidecar/realtime";
+import { type RealtimeDiagnostics, voiceExchangeActive } from "@sidecar/realtime";
 import {
   isSpeechOffer,
   isSpeechOutcome,
@@ -96,7 +96,6 @@ import {
   type VoiceCommand,
   type VoiceCommandOutcome,
   type VoiceView,
-  voiceExchangeActive,
 } from "./messages/voice-view";
 
 export interface WireGuard<Value> {
